2-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-server"
version = "1.0.0"
description = "Local MCP server with Ollama integration — clone-and-run"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "anyio>=4.0.0",
    "mcp>=1.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.2.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "streamlit>=1.30.0",
    "keyring>=24.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
mcp-server = "mcp_server.main:main"

[tool.hatch.build.targets.wheel]
packages = ["mcp_server"]

# ---------------------------------------------------------------------------
# Ruff
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
